# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UTP TIS - Ecommerce API",
        "description": "API sederhana ecommerce menggunakan Laravel dengan JSON storage. Dibuat untuk UTP TIS.",
        "contact": {
            "name": "Naila Keisha",
            "email": "naila@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Tampilkan semua barang",
                "operationId": "2cbf07eb95d033f8775c36424ac8fa06",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil semua data barang",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil mengambil semua data barang"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama_barang": {
                                                        "type": "string",
                                                        "example": "Laptop Asus"
                                                    },
                                                    "harga": {
                                                        "type": "integer",
                                                        "example": 8500000
                                                    },
                                                    "stok": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "deskripsi": {
                                                        "type": "string",
                                                        "example": "Laptop gaming"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2026-04-19 10:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2026-04-19 10:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Tambah barang baru",
                "operationId": "af37857663e6805dd367a55f374c9ee1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_barang",
                                    "harga"
                                ],
                                "properties": {
                                    "nama_barang": {
                                        "type": "string",
                                        "example": "Laptop Asus"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 8500000
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Laptop gaming terjangkau"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Barang berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Tampilkan barang berdasarkan ID",
                "operationId": "7f1ffbd9fa2b704c61ebcccb636ebb27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data barang",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil mengambil data barang"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item dengan ID 99 tidak Ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Update seluruh data barang",
                "operationId": "4ac9fb55d8a18bd1cba3a29d84f9fd4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_barang",
                                    "harga",
                                    "stok"
                                ],
                                "properties": {
                                    "nama_barang": {
                                        "type": "string",
                                        "example": "Laptop Asus Pro"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 9000000
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Versi terbaru"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barang berhasil diupdate sepenuhnya"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Hapus barang",
                "operationId": "0f88b60b73861d5414f8ea2fedd3dfc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barang berhasil dihapus"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Items"
                ],
                "summary": "Update sebagian data barang",
                "operationId": "8171e1e0bd396fd89495538eb21a8869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_barang": {
                                        "type": "string",
                                        "example": "Laptop Baru"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 7500000
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Update deskripsi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barang berhasil diupdate sebagian"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Items",
            "description": "Endpoint untuk manajemen barang"
        }
    ]
}